# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "bale-broadcast-ai API",
    "version": "1.0.0",
    "description": "API واحد و نسخه‌بندی‌شده‌ی پلتفرم پیام‌رسانی/فروش هوشمند روی بله.\n\n## احراز هویت\nهمه‌ی endpoint ها (به‌جز /auth/register و /auth/login) به هدر Authorization: Bearer <token> نیاز دارند. دو نوع توکن پذیرفته می‌شود و هر دو از یک هدر یکسان استفاده می‌کنند:\n\n1. JWT نشست داشبورد — از POST /auth/login یا POST /auth/register گرفته می‌شود. اعتبار ۷ روزه دارد.\n   مثال: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n2. کلید API بلندمدت — از POST /api-keys ساخته می‌شود، با پیشوند abai_live_ شروع می‌شود.\n   مثال: Authorization: Bearer abai_live_9f6e817cb2a1...\n\nسرور نوع توکن را خودکار از روی پیشوندش تشخیص می‌دهد. هر دو userId احراز هویت‌شده را تعیین می‌کنند و پلتفرم داده‌ی هر کاربر را کاملاً از بقیه ایزوله نگه می‌دارد.\n\n## کدهای خطای استاندارد\n- 401: توکن ارسال نشده، نامعتبر، منقضی یا باطل‌شده است.\n- 403: توکن معتبر است ولی اجازه‌ی دسترسی به این بخش را ندارد (مثلاً /admin/* فقط برای owner).\n\nهر پاسخ خطا یک شیء ساده با فیلد error (رشته‌ی قابل‌نمایش) برمی‌گرداند."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT یا API Key",
        "description": "یا JWT گرفته‌شده از /auth/login و /auth/register، یا یک کلید API که با POST /api-keys ساخته‌اید (پیشوند abai_live_). هر دو در همین یک هدر Authorization: Bearer <token> فرستاده می‌شوند."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "درخواست نامعتبر — فیلد الزامی خالی است یا مقدار داده‌شده نادرست است.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "فیلدهای name و descriptionText الزامی هستند." } } }
      },
      "Unauthorized": {
        "description": "توکن ارسال نشده، نامعتبر، منقضی یا باطل‌شده است.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "توکن احراز هویت ارسال نشده (Authorization: Bearer <token>)." } } }
      },
...
      "NotFound": {
        "description": "منبع یافت نشد — یا اصلاً وجود ندارد، یا متعلق به کاربر دیگری است (این دو حالت عمداً یکسان جواب داده می‌شوند تا اطلاعاتی درباره‌ی داده‌ی کاربران دیگر درز نکند).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "محصول یافت نشد." } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string", "description": "پیام خطای قابل‌نمایش به کاربر" } },
        "required": ["error"]
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "shortCode": { "type": "string" },
          "descriptionText": { "type": "string" },
          "isActive": { "type": "boolean" },
          "userId": { "type": "string", "nullable": true, "description": "null یعنی محصول سراسری (global) و همه‌ی کاربران می‌بینندش" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Contact": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "phone": { "type": "string", "example": "989123456789" },
          "status": { "type": "string", "enum": ["pending", "sent", "failed", "no_bale"] },
          "sentAt": { "type": "string", "format": "date-time", "nullable": true },
          "errorMessage": { "type": "string", "nullable": true }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sender": { "type": "string", "enum": ["user", "ai", "human"] },
          "text": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "contactId": { "type": "string" },
          "productId": { "type": "string", "nullable": true },
          "state": { "type": "string", "enum": ["active", "completed", "abandoned"] },
          "needsHuman": { "type": "boolean" },
          "escalatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "escalationReason": { "type": "string", "nullable": true },
          "contact": { "$ref": "#/components/schemas/Contact" },
          "product": { "$ref": "#/components/schemas/Product" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" } }
        }
      },
      "WaitingConversation": {
        "allOf": [
          { "$ref": "#/components/schemas/Conversation" },
          { "type": "object", "properties": { "waitingMinutes": { "type": "integer", "nullable": true, "description": "چند دقیقه از ارجاع می‌گذرد" } } }
        ]
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/products": {
      "get": {
        "summary": "لیست محصولات من + محصولات سراسری",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "ساخت محصول جدید",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["name", "descriptionText"], "properties": { "name": { "type": "string" }, "shortCode": { "type": "string" }, "descriptionText": { "type": "string" } } },
              "example": { "name": "کرم مرطوب‌کننده", "descriptionText": "حجم ۵۰ میلی‌لیتر، قیمت ۲۵۰ هزار تومان" }
            }
          }
        },
        "responses": {
          "201": { "description": "ساخته شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "جزئیات یک محصول",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "ویرایش/فعال-غیرفعال کردن محصول (فقط مالک محصول، نه محصولات سراسری)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" }, "descriptionText": { "type": "string" }, "isActive": { "type": "boolean" } } },
              "example": { "isActive": false }
            }
          }
        },
        "responses": {
          "200": { "description": "به‌روزرسانی شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/conversations/active": {
      "get": {
        "summary": "مکالمات فعال (در حال گفتگوی خودکار با AI)",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Conversation" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/conversations/waiting": {
      "get": {
        "summary": "مکالمات منتظر پاسخ من (ارجاع‌شده به انسان)",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/WaitingConversation" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/conversations/{id}": {
      "get": {
        "summary": "تاریخچه کامل یک مکالمه",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "conversation": { "$ref": "#/components/schemas/Conversation" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/conversations/{id}/reply": {
      "post": {
        "summary": "پاسخ دستی به یک مکالمه‌ی ارجاع‌شده — مستقیماً به بازو می‌رود",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string" } } }, "example": { "text": "سلام، بله موجوده. ارسالش رایگانه." } } }
        },
        "responses": {
          "200": { "description": "ارسال شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
          "400": { "description": "متن خالی است یا مخاطب chat_id بازو ندارد", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/conversations/{id}/resolve": {
      "post": {
        "summary": "پایان ارجاع — مکالمه به AI برمی‌گردد و دوباره خودکار پاسخ می‌دهد",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "conversation": { "$ref": "#/components/schemas/Conversation" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}